Testing for empty courses run state
